requirements.txt
discord.py==1.7.2
discord-ext-menus==1.1
pyyaml==5.4.1
Jishaku==1.20.0.220
aiomysql==0.0.21